pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "taxweave-atlas"
version = "0.1.0"
description = "Synthetic US individual tax dataset generator (local, foundation scaffold)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6",
  "PyYAML>=6.0",
  "click>=8.1",
]

[project.optional-dependencies]
dev = ["ruff>=0.3"]

[project.scripts]
taxweave-atlas = "taxweave_atlas.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
